+++ controller/chimera_page_controller.js
@@ -3,13 +3,12 @@
  *
  * @author        Jelle De Loecker   <jelle@develry.be>
  * @since         0.2.0
- * @version       0.3.0
+ * @version       0.5.0
  */
-var Page = Function.inherits('Alchemy.EditorChimeraController', function PageChimeraController(conduit, options) {
-	PageChimeraController.super.call(this, conduit, options);
+var Page = Function.inherits('Alchemy.Controller.Chimera.Editor', function Page(conduit, options) {
+	Page.super.call(this, conduit, options);
 });
 
-Page.setMethod(function edit(conduit) {
-
+Page.setAction(function edit(conduit) {
 	this.render('chimera/page/edit');
 });