+++ controller/chimera_static_controller.js
@@ -5,8 +5,8 @@
  * @since         0.2.0
  * @version       0.3.0
  */
-var ChimeraStatic = Function.inherits('Alchemy.ChimeraController', function ChimeraStaticController(conduit) {
-	ChimeraStaticController.super.call(this, conduit);
+var ChimeraStatic = Function.inherits('Alchemy.Controller.Chimera', function Static(conduit) {
+	Static.super.call(this, conduit);
 });
 
 /**
@@ -14,11 +14,11 @@ var ChimeraStatic = Function.inherits('Alchemy.ChimeraController', function Chim
  *
  * @param   {Conduit}   conduit
  */
-ChimeraStatic.setMethod(function dashboard(conduit) {
+ChimeraStatic.setAction(function dashboard(conduit) {
 	this.set('pagetitle', 'Dashboard');
 	this.render('chimera/dashboard');
 });
 
-ChimeraStatic.setMethod(function pageEditor(conduit) {
+ChimeraStatic.setAction(function pageEditor(conduit) {
 	this.render('chimera/page_editor');
 });